+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.1.7] - 2022-12-10
+### Added
+- Recipes for large, huge, and super loginar urns
 ### Fixed
 - Some urn recipes not working correctly (failing to upgrade, changing gem color unintentionally) [#1]